requirements.txt
fastapi
uvicorn
python-multipart
requests
pydantic==2.7.0
python-jose
passlib
bcrypt==3.2.2
azure-ai-documentintelligence==1.0.0
azure-core>=1.30.0
prisma
